Refactor structure for formulations
