+++ cooper/__init__.py
@@ -15,11 +15,13 @@ except PackageNotFoundError:
 
     warnings.warn("Could not retrieve cooper version!")
 
-from cooper.augmented_lagrangian import AugmentedLagrangianFormulation
 from cooper.constrained_optimizer import ConstrainedOptimizer
-from cooper.formulation import Formulation, UnconstrainedFormulation
-from cooper.lagrangian_formulation import LagrangianFormulation
+from cooper.formulation import (
+    Formulation,
+    LagrangianFormulation,
+    UnconstrainedFormulation,
+)
 from cooper.problem import CMPState, ConstrainedMinimizationProblem
 from cooper.state_logger import StateLogger
 
-from . import optim, utils
+from . import formulation, optim, utils
