+++ cooper/problem.py
@@ -1,6 +1,6 @@
 import abc
 from dataclasses import dataclass
-from typing import Any, Callable, Dict, List, Optional
+from typing import Optional
 
 import torch
